+++ views/pages/login.ejs
@@ -4,7 +4,7 @@
 
     <%= messages.error %>
         <% }; %>
-<form action="/user/loginForm" method="POST" id="login">
+<form action="/login" method="POST" id="login">
 
     <h1>Login</h1>
