+++ views/pages/register.ejs
@@ -9,7 +9,7 @@
     </div>
 </div>
 
-<form action="/user/registerForm" method="POST" id="register">
+<form action="/register" method="POST" id="register">
 
     <h1>Sign Up</h1>
